manually fix remaining clippy lints
