+++ reflectapi-demo/src/main.rs
@@ -169,7 +169,8 @@ async fn pets_list(
     let result_items = pets
         .iter()
         .skip(cursor)
-        .take(limit).cloned()
+        .take(limit)
+        .cloned()
         .collect::<Vec<_>>();
     let result_cursor = if result_items.is_empty() {
         None
